+++ src/game/main.js
@@ -16,6 +16,7 @@ import { CollectionScene } from './scenes/CollectionScene'
 import { StatsScene } from './scenes/StatsScene'
 import { MapScene } from './scenes/MapScene'
 import { CreditsScene } from './scenes/CreditsScene'
+import { LicensesScene } from './scenes/LicensesScene'
 import { AUTO, Scale, Game } from 'phaser'
 import { GAME_WIDTH, GAME_HEIGHT } from './config/gameConfig'
 
@@ -49,6 +50,7 @@ const config = {
     StatsScene,
     MapScene,
     CreditsScene,
+    LicensesScene,
   ],
 }